📝 (content): enhance social technologies category with detailed insights
The social technologies category content is expanded to provide a deeper
understanding of the philosophies and principles that underpin these
frameworks. This includes new sections on self-organisation, empirical
inquiry, and collective intelligence, as well as guidance on cultivating
a social technology ethos. These additions aim to better inform readers
about the importance of adaptability and continuous learning in complex
environments.

💡 (layouts): improve category and tag list layouts for better UX

The category and tag list layouts are refined to enhance user
experience. The changes include updating button links for better
navigation and adding a headline section to the tags list for improved
contextual information. These updates aim to make the site more
intuitive and informative for users.

🔧 (css): update styles for two-column text layout

The CSS for the two-column text layout is updated to ensure better
readability and presentation. Headings and paragraphs are adjusted to
avoid awkward breaks, improving the overall visual flow of the content.
This change enhances the aesthetic and readability of the site content.