📝 (layouts): update HTML structure for better readability and functionality
Remove the inline style limiting the container width to 900px to allow
for a more responsive design. Add an "Overview" heading to improve
content structure and readability. Duplicate the share-bar and add a
comments section to enhance user interaction and engagement.

In the admin-dropdown component, conditionally render the ResourceId
button only for resources, and correct the permalink URL to use the
dynamic $url variable for accuracy. These changes improve the user
experience and maintainability of the site.